// src/modules/users/enums/user-role.enum.ts
export enum UserRoleEnum {
  BUYER = 'buyer',
  SELLER = 'seller',
  MANAGER = 'manager',
  ADMIN = 'admin',
  MECHANIC = 'mechanic',
  DEALERSHIP_ADMIN = 'dealership-admin',
  DEALERSHIP_MANAGER = 'dealership-manager',
  DEALERSHIP_SELLER = 'dealership-seller',
  DEALERSHIP_MECHANIC = 'dealership-mechanic',
}